requirements.txt
discord.py==2.3.2
openai==0.27.8
python-dotenv==1.0.0
aiohttp==3.9.0b0
